requirements.txt
# ── AI & ML Core ────────────────────────────────────────────────────────────
numpy>=1.26.0
opencv-python>=4.8.0
deepface>=0.0.92
tf-keras>=2.15.0
mediapipe>=0.10.0
pyyaml>=6.0
Pillow>=10.0.0

# ── Blockchain Simulation ──────────────────────────────────────────────────
# (Built-in hashlib and json are used)
imagehash>=4.3.1
scikit-image>=0.21.0
scipy>=1.11.0
